+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ADDED:
 
+- Possibility to retrieve the connection from query builders
+
 CHANGED:
 
 FIXED:
+
+- Parentheses are added for the ON clause
 
 ## 0.1.0 - 2017-01-29